pyproject.toml
[project]
name = "agentic-day-planner"
version = "0.1.0"
description = "Self-hosted agentic day planner with sandboxed file access"
authors = [
    {name = "Rishabh Sharma", email = "99526730+sharma-rishabh@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49",
    "python-telegram-bot>=21",
    "python-dotenv>=1.1",
    "pyyaml>=6.0",
    "matplotlib (>=3.10.8,<4.0.0)",
    "mempalace (>=3.3.1,<4.0.0)",
    "chromadb (>=1.5.8,<2.0.0)",
]

[project.scripts]
planner = "planner_agent.main:run"

[tool.poetry]
packages = [{include = "planner_agent", from = "src"}]

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
pytest-asyncio = ">=0.25"
ruff = ">=0.11"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
